requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic[email]==2.9.0
requests==2.32.0
mangum
google-generativeai